+++ package.json
@@ -26,7 +26,6 @@
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-prettier": "^4.0.0",
     "express": "^4.17.1",
-    "jsonwebtoken": "^8.5.1",
     "mysql2": "^2.3.2",
     "nodemon": "^2.0.14",
     "prettier": "^2.4.1",
